# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog Project API",
        "description": "API documentation for the blog system.",
        "contact": {
            "email": "farzane.rahmani2@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get authenticated user's posts",
                "operationId": "fb84b8ac53d4f3c7be8b9d99ccdfaaf4",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Hello world"
                                                    },
                                                    "body": {
                                                        "type": "string",
                                                        "example": "Lorem Ipsum is simply dummy text of the printing and typesetting industry..."
                                                    },
                                                    "views_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "author": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Sara"
                                                            },
                                                            "profile_image": {
                                                                "type": "string",
                                                                "example": "http://localhost/storage/profile/sara.jpg",
                                                                "nullable": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2025-05-24 12:34:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get a single post by ID",
                "operationId": "b76e6fb47b679ab1136d8a0a5101ac9f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Qui qui ea dolores sint."
                                                },
                                                "body": {
                                                    "type": "string",
                                                    "example": "Labore et hic dolorum..."
                                                },
                                                "views_count": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "author": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Tara Pagac"
                                                        },
                                                        "profile_image": {
                                                            "type": "string",
                                                            "example": "http://localhost:8000/storage/profile_images/ukzz2WRY2zouwJd72zFJJzLNBoLyww1jKDkxawll.jpg",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "server_time": {
                                            "type": "string",
                                            "example": "2025-05-24 16:35:01"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}